pyproject.toml
[project]
name = "lbg-customer-ai-super-agent"
version = "0.1.0"
description = "LBG Coaching Agent MVP - Phase 1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.39.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "faiss-cpu>=1.8.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.8.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["coaching_agent", "data", "api", "demo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
